# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js Blogg Application API",
    "description": "Yet Another Node.js Blogg Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Lists",
      "description": "API for users in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/bucketlist": {
      "post": {
        "tags": [
          "List"
        ],
        "description": "Create new list",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "List to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/List"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New list is created",
            "content": [
              "application/json:"
            ],
            "schema": {
              "$ref": "#/definitions/List"
            }
          }
        }
      },
      "get": {
        "tags": [
          "List"
        ],
        "summary": "Get all lists in system",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "produces": [
              "application/json:"
            ],
            "schema": {
              "$ref": "#/definitions/List"
            }
          }
        }
      }
    },
    "/bucketlist/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of list that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete list with given ID",
        "tags": [
          "List"
        ],
        "responses": {
          "200": {
            "description": "List is deleted",
            "schema": {
              "$ref": "#/definitions/List"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "List": {
      "type": "object",
      "required": [
        "title",
        "category"
      ],
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "type": "string",
          "enum": ["High", "Medium", "Low"]
        }
      }
    }
  }
}